apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flasgger>=0.9.7.1",
    "flask>=3.1.3",
    "python-dotenv>=1.2.2",
]
